Added Compatibility for Directional Lights in the Level Transitions.
Level Transitions can now accept Directional Lights for a prefab. Currently, there are no circumstances in what the skybox changes into, but those can be added as we make new levels.

Directional Lights must be child objects to an empty Game Object prefab in order for the script to work correctly.