Updated Main Menu (See Changes Below)
List of Changes:

Visual Button Layout:
Updated the Sprites used for the Button Layout on the Main Menu (the right arrow has been changed to a checkmark to better illustrate its purpose in menu navigation).

Added sprites for left button inputs, and right button inputs during a change in selection. These are included to show that they have no effect.

Menu:
The Menu now exists in "World Space." this means that how the menu appears and moves should remain consistent regardless of the screen size it is being viewed on.

Added a "sliding" effect when changing selection between menu items. This is programmed to be delay the ability to input items until the selection is complete (in other words, you cannot change selection or confirm a selected menu item until the animation is finished).
You can no longer change the selection of menu items while confirming a selected menu item (which carried over the confirmation time into the next selection).