Implemented "Settings" Option in Main Menu
The "Settings" option on the Main Menu now has an actual function. When selected, the menu will shift to a new set of options: "Master Volume," "Music Volume," and "Sound Effect Volume." As their names imply, these are intended to control the volume at which sound plays (although this currently DOES NOT have any actual effect on volume, the script contains static variables for each of these values, so all that would need to be added is the code for the sound to actually be affected by these changes). These values can be adjusted by pressing the left or right buttons on the controller: when over an adjustable option, the button's visual indicator will change to reflect the change in context for the button's functionality.

The Credits screen also has had the visual indicator change to better match the actual usability of the buttons (in the case of the credits screen, only the right button can be used).