Implemented Pause Menu, and Possibly Button Controls
When the game pauses, a menu now appears, giving you several options:

Resume: Closes the Menu and unpauses the game.
Settings: Allows for adjustment of sound volume.
Quit: Asks for confirmation from the player to close the application. If the player selects "Yes," the application closes. If the player selects "No," it goes back to the beginning of the pause menu.

Additions have been made to these scripts that seem to allow for button inputs from the VZController, but I am unable to test for this, so I also still allow for the menus to be controlled with the arrow keys (this also adjusts the camera in the player's view, but once the player is actually wearing the headset, and those two different functions are bound to separate actions, this shouldn't be an issue).