Added Background Image to the main menu +music
added a background image to the main menu, it should be treated as a placeholder since the VZ controller is being used with means the player can move their head if they so choose.

I also added in the main menu theme.

BUGS: When the game starts the "Press to continue text disappears" Unknown as to why.